backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.32.1
sqlalchemy==2.0.37
asyncpg==0.30.0
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
python-dotenv==1.0.1
email-validator==2.2.0
aiosqlite==0.20.0
# Migrations
alembic==1.18.4
# Rate limiting
slowapi==0.1.9
# Error tracking
sentry-sdk[fastapi]==2.58.0
# Testing
pytest==8.3.5
pytest-asyncio==0.24.0
